+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -337,7 +337,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kezxqgbzrp6wwpv7s4gys90p"),
+  _id: z.string().default("reasoning_01kf08b1jfsr1dz52ymqd9r7pb"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -362,7 +362,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kezxqgbzrp6wwpv7s4gys90p"),
+  id: z.string().default("reasoning_01kf08b1jfsr1dz52ymqd9r7pb"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),
@@ -510,7 +510,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kezxqgbva58jpqg7btdv9g5j"),
+  _id: z.string().default("reasoning_01kf08b1j3ky4kjsg2m5cwt41f"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),
@@ -535,7 +535,7 @@ export const ReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kezxqgbva58jpqg7btdv9g5j"),
+  id: z.string().default("reasoning_01kf08b1j3ky4kjsg2m5cwt41f"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$outboundSchema,
   reasoning: z.string(),