+++ packages/orq-rc/RELEASES.md
@@ -5839,3 +5839,13 @@ Based on:
 - [typescript v4.2.0-rc.30] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.30 - packages/orq-rc
+
+## 2026-01-15 07:21:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.3 (2.795.8) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.31] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.31] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.31 - packages/orq-rc